+++ BoxDemo/PreviewController.m
@@ -27,7 +27,7 @@
     IBOutlet UILabel *labSizeY;
     IBOutlet UILabel *labSizeZ;
     
-    float width, height, length;
+    CGFloat width, height, length;
 }
 
 @property (nonatomic, weak) SCNNode *theCamera;
@@ -36,7 +36,7 @@
 
 @implementation PreviewController
 
-+(NSArray *)getFilenamelistOfType:(NSString *)type fromDirPath:(NSString *)dirPath
++ (NSArray *)getFilenamelistOfType:(NSString *)type fromDirPath:(NSString *)dirPath
 {
     NSMutableArray *filenamelist = [NSMutableArray arrayWithCapacity:10];
     NSArray *tmplist = [[NSFileManager defaultManager] contentsOfDirectoryAtPath:dirPath error:nil];
@@ -53,13 +53,13 @@
     return filenamelist;
 }
 
-+(BOOL)isFileExistAtPath:(NSString*)fileFullPath {
++ (BOOL)isFileExistAtPath:(NSString *)fileFullPath {
     BOOL isExist = NO;
     isExist = [[NSFileManager defaultManager] fileExistsAtPath:fileFullPath];
     return isExist;
 }
 
--(void)enumPath:(NSString *)homePath
+- (void)enumPath:(NSString *)homePath
 {
     NSFileManager *fm = [NSFileManager defaultManager];
     NSDirectoryEnumerator *dirEnum = [fm enumeratorAtPath:homePath];
@@ -153,7 +153,7 @@
     self.scnView.pointOfView.transform = SCNMatrix4Mult(mat4_t, mat4_r);
 }
 
--(void)initCylinder:(SCNScene *)scene
+- (void)initCylinder:(SCNScene *)scene
 {
     // board geometry
     SCNGeometry *boardGeometry = [SCNCylinder cylinderWithRadius:width/2 height:1.0f];
@@ -175,7 +175,7 @@
     [scene.rootNode addChildNode:boxNode];
 }
 
--(void)initCubeBox:(SCNScene *)scene
+- (void)initCubeBox:(SCNScene *)scene
 {
     // board geometry
     SCNBox *boardGeometry = [SCNBox boxWithWidth:width height:height length:1.0f chamferRadius:0];//width = x, height = y, length = z;
@@ -195,7 +195,7 @@
     [scene.rootNode addChildNode:boxNode];
     
     //direction geometry
-    float dirWidth = width/3;
+    CGFloat dirWidth = width/3;
     for (int i = 0; i < 3; i++) {
         SCNBox *box = [SCNBox boxWithWidth:dirWidth height:2 length:2 chamferRadius:0];
         SCNNode *boxNode = [SCNNode nodeWithGeometry:box];
@@ -221,7 +221,7 @@
     }
 }
 
--(void)refreshSize:(SCNVector3)vector
+- (void)refreshSize:(SCNVector3)vector
 {
     SCNVector3 v3Min, v3Max;
     [_theModelNode getBoundingBoxMin:&v3Min max:&v3Max];
@@ -231,7 +231,7 @@
     labSizeZ.text = [NSString stringWithFormat:@"SizeZ:%0.2fmm", fabsf(v3Max.z-v3Min.z)*vector.z];
 }
 
--(void)printMatrix4:(SCNMatrix4)mat4
+- (void)printMatrix4:(SCNMatrix4)mat4
 {
     printf("%f, %f, %f, %f \n%f, %f, %f, %f \n%f, %f, %f, %f \n%f, %f, %f, %f\n\n\n",
            mat4.m11, mat4.m12, mat4.m13, mat4.m14,
@@ -240,7 +240,7 @@
            mat4.m41, mat4.m42, mat4.m43, mat4.m44);
 }
 
--(IBAction)doCloseClick:(id)sender
+- (IBAction)doCloseClick:(id)sender
 {
     [self dismissViewControllerAnimated:YES completion:nil];
 }
